requirements.txt
Jinja2
pandas
markdown
panel
perspective-python
setuptools
ydata-profiling
